разделил функцию создания массивов на модули
