+++ js/main.js
@@ -1,52 +1,6 @@
-const PHOTO_COUNT = 25;
+import {PHOTO_COUNT, DESCRIPTIONS} from './util.js';
 
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'У моего кота получилась фотография лучше.',
-];
-
-const DESCRIPTIONS = [
-  'Утро!',
-  'Котик',
-  'Солнышко',
-  'Красивое фото',
-  'Что-то новенькое',
-];
-
-const NAMES = [
-  'Иван',
-  'Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-];
-
-const getRandomIntInclusive = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1) + min);
-};
-
-const counter = () => {
-  let sum = 0;
-  return () => {
-    sum = sum + 1;
-    return sum;
-  };
-};
-
-const uniquePhoto = counter();
-const uniqueId = counter();
-const getRandomArrayElement = (elements) => elements[getRandomIntInclusive(0, elements.length - 1)];
-
-const createComments = () => ({
-  id: getRandomIntInclusive(0, 30),
-  avatar: `img/avatar-${getRandomIntInclusive(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES),
-});
+import {getRandomArrayElement, getRandomIntInclusive, uniquePhoto, uniqueId, createComments} from './script.js';
 
 
 const createPhoto = () => {
