+++ js/util.js
@@ -1,0 +1,26 @@
+const PHOTO_COUNT = 25;
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'У моего кота получилась фотография лучше.',
+];
+
+const DESCRIPTIONS = [
+  'Утро!',
+  'Котик',
+  'Солнышко',
+  'Красивое фото',
+  'Что-то новенькое',
+];
+
+const NAMES = [
+  'Иван',
+  'Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+];
+
+export {PHOTO_COUNT, MESSAGES, DESCRIPTIONS, NAMES};
